% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{function.portal}
\alias{function.portal}
\title{Taxonomy of species sampled}
\format{
A data frame with X number of columns and Y number of rows, the names of the columns are: "label","species_id","genus","species", "taxa","ott_name","approximate_match","ott_id","is_synonym", and "flags"
}
\source{
\url{https://lunasare.github.io/spring2023-data-science/data/portal-species-taxonomy.csv}
}
\usage{
function.portal
}
\description{
A data set containing the taxonomy of species sampled for the Portal Project Data.
We followed the taxonomy from the Open Tree of Life .
}
\keyword{datasets}
